+++ frontend/src/features/accounts/manage-accounts.tsx
@@ -3,8 +3,8 @@
 import { Account } from '@/../../infrastructure/graphql/api/codegen/appsync';
 import { Button } from '@/components/ui/button';
 import { useNewAccount } from '@/hooks/use-new-account';
-import { DataTable } from '@/features/accounts/data-table';
-import { columns } from '@/features/accounts/columns';
+import { DataTable } from '@/components/data-table';
+import { columns } from '@/app/(main)/accounts/columns';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Plus } from 'lucide-react';
