+++ frontend/src/providers/sheet-provider.tsx
@@ -2,8 +2,8 @@
 
 import { useMountedState } from 'react-use';
 
-import EditAccountSheet from '@/features/accounts/edit-account-sheet';
-import NewAccountSheet from '@/features/accounts/new-account-sheet';
+import EditAccountSheet from '@/features/accounts/components/edit-account-sheet';
+import NewAccountSheet from '@/features/accounts/components/new-account-sheet';
 
 export const SheetProvider = () => {
   const isMounted = useMountedState();
